fix(fill): 3d honeycomb density must multiply in f64 like upstream
Fill.cpp:1283 computes float(0.01 * params.density) — the double
literal promotes the float percent to f64 before the product, giving
f32(0.15000000596) at 15%. The ares dispatch multiplied in pure f32
(f32(0.01)×15 → 0.14999999403, two ulps low), shifting gridSize
(3579535.68 vs Orca 3579535.104) and every critical product by
~one unit. Match the sibling fill entities' (0.01_f64 *
f64::from(..)) as f32 form; the fixture diff drops 753 -> 323
lines.

+++ crates/ares-core/src/project_slice/fill_entities/three_d_honeycomb.rs
@@ -18,7 +18,7 @@ pub(super) fn append(
         spacing: fill.params.spacing,
         overlap: fill.params.overlap,
         angle: fill.params.angle,
-        density: 0.01 * fill.params.density,
+        density: (0.01_f64 * f64::from(fill.params.density)) as f32,
         multiline: fill.params.multiline,
         anchor_length: fill.params.anchor_length,
         anchor_length_max: fill.params.anchor_length_max,
